Change route names and file names (#34)
* moved welcome/index to room/index and changed references

* UNTESTED changed scene1 view to home and changed dependencies in controllers and routes accordingly

* UNTESTED changed scene1 view to home and changed dependencies in controllers and routes accordingly

* fixed wonky names of files and dependencies
