--- app/controllers/video_controller.rb
@@ -1,4 +1,0 @@
-class VideoController < ApplicationController
-  def index
-  end
-end
