+++ app/controllers/welcome_controller.rb
@@ -1,33 +1,17 @@
 class WelcomeController < ApplicationController
-
-  def is_authenticated
-    return session[:user_id] != nil
-  end
-
   def from_outside_controller
     puts 'OUTSIDE CONTROLLER'
     if is_authenticated()
-      render json: {location: '/welcome/index'}
+      puts 'rendering room/index'
+      render json: {location: '/room/index'}
     else
-      render json: {location: '/welcome/scene1'}
-    end
-  end
-
-  def index
-    puts session[:user_id]
-    if is_authenticated()
-      @roomID = session[:room_id]
-      @roomName = session[:room_name]
-      @userID = session[:user_id]
-      puts 'hello'
-      render '/welcome/index'
-    else
-      redirect_to '/welcome/scene1'
+      puts 'rendering welcome/home'
+      render json: {location: '/welcome/home'}
     end
   end
 
   def redirect
-    redirect_to '/welcome/scene1'
+    redirect_to '/welcome/home'
   end
 
   def scene1
