--- app/helpers/video_helper.rb
@@ -1,2 +1,0 @@
-module VideoHelper
-end
