+++ app/views/room/index.html.erb
