--- test/controllers/video_controller_test.rb
@@ -1,9 +1,0 @@
-require 'test_helper'
-
-class VideoControllerTest < ActionDispatch::IntegrationTest
-  test "should get index" do
-    get video_index_url
-    assert_response :success
-  end
-
-end
